Key generator functionality was separated to it's own class. Tests was refactored
